refactor(secondary-screen): extract pure Now Playing helpers to top-level functions
Move the pure presentation helpers the Now Playing panel relies on out of
`_SecondaryScreenState` into `secondary_screen/now_playing_helpers.dart` as
top-level functions: `resolveTheme`, `contrastRatio`, `panelScheme`,
`buildNowPlayingBoxart`, `buildNowPlayingStat`, `formatPlayTime`,
`formatLastPlayed`.

Each is a pure mapping from its inputs (no widget-state reads), so the bodies
move byte-identically with only the private->public rename. The state-coupled
`_formatSessionTime` (reads `_sessionWatch`) and the panel shell itself stay in
the host. All call sites updated (`build`, the Now Playing panel, dock and
launcher all resolve the panel scheme via the shared `panelScheme`); the
now-unused `theme_provider` host import is dropped.

Also relocates a stray doc paragraph: the "Renders the Now Playing page"
comment was stacked above `_resolveTheme` but describes `_buildNowPlayingPanel`
— it now sits directly above the panel it documents (no text lost, no runtime
effect). Host 2332 -> 2209 lines.

Second of the Phase-2 secondary_screen extractions; behaviour unchanged. Sets
up a later clean Now Playing panel widget (which needs the nested app dock
extracted first).
